invert colors and move components
